Crate rome_console

Modules

fmt

Macros

markup

Structs

BufferConsole

Implementation of Console storing all printed messages to a memory buffer

EnvConsole

Implementation of Console printing messages to the standard output and standard error

HorizontalLine

A horizontal line with the given print width

Markup

Root type returned by the markup macro: this is simply a container for a list of markup nodes

MarkupBuf

MarkupNode

Implementation of a single “markup node”: a piece of text with a number of associated styles applied to it

Message

Individual message entry printed to a BufferConsole

Enums

LogLevel

Determines the “ouput stream” a message should get printed to

MarkupElement

Enumeration of all the supported markup elements

Traits

Console

Generic abstraction over printing markup and diagnostics to an output, which can be a terminal, a file, a memory buffer …

ConsoleExt

Extension trait for Console providing convenience printing methods